Refactor integration tests
Now that a lot of things in the build process of a
u-bmc image has changed the integration tests needed
to be modified to account for these changes.
They still got a few issues but the general layout should
be final now.
