+++ integration/testcmd/acme/uinit/test.go
@@ -10,18 +10,18 @@ import (
 	"time"
 
 	"github.com/u-root/u-bmc/config"
-	"github.com/u-root/u-bmc/integration/utils"
+	"github.com/u-root/u-bmc/integration/util"
 	"github.com/u-root/u-bmc/pkg/bmc"
 	"github.com/u-root/u-bmc/pkg/bmc/ttime"
-	"github.com/u-root/u-bmc/platform/quanta-f06-leopard-ddr3/pkg/platform"
+	"github.com/u-root/u-bmc/platform/qemu-virt-a72/pkg/platform"
 )
 
 func uinit() error {
 	p := platform.Platform()
 	defer p.Close()
 
-	ca := utils.NewTestCA()
-	rt := utils.NewTestRoughtimeServer()
+	ca := util.NewTestCA()
+	rt := util.NewTestRoughtimeServer()
 
 	c := config.DefaultConfig
 	c.RoughtimeServers = []ttime.RoughtimeServer{rt.Config}
@@ -55,8 +55,8 @@ func uinit() error {
 
 func main() {
 	if err := uinit(); err != nil {
-		utils.FailTest(err)
+		util.FailTest(err)
 	} else {
-		utils.PassTest()
+		util.PassTest()
 	}
 }
