+++ integration/util/network.go
@@ -1,0 +1,48 @@
+// Copyright 2018 the u-root Authors. All rights reserved
+// Use of this source code is governed by a BSD-style
+// license that can be found in the LICENSE file.
+
+package util
+
+import (
+	"fmt"
+
+	"github.com/vishvananda/netlink"
+	"golang.org/x/sys/unix"
+)
+
+func AddIP(cidr string, iface string) error {
+	l, err := netlink.LinkByName(iface)
+	if err != nil {
+		return fmt.Errorf("unable to get interface %s: %v", iface, err)
+	}
+	addr, err := netlink.ParseAddr(cidr)
+	if err != nil {
+		return fmt.Errorf("netlink.ParseAddr %v: %v", cidr, err)
+	}
+	h, err := netlink.NewHandle(unix.NETLINK_ROUTE)
+	if err != nil {
+		return fmt.Errorf("netlink.NewHandle: %v", err)
+	}
+	defer h.Delete()
+	if err := h.AddrReplace(l, addr); err != nil {
+		return fmt.Errorf("AddrReplace(%v): %v", addr, err)
+	}
+	return nil
+}
+
+func SetLinkUp(iface string) error {
+	l, err := netlink.LinkByName(iface)
+	if err != nil {
+		return fmt.Errorf("unable to get interface %s: %v", iface, err)
+	}
+	h, err := netlink.NewHandle(unix.NETLINK_ROUTE)
+	if err != nil {
+		return fmt.Errorf("netlink.NewHandle: %v", err)
+	}
+	defer h.Delete()
+	if err := h.LinkSetUp(l); err != nil {
+		return fmt.Errorf("handle.LinkSetUp: %v", err)
+	}
+	return nil
+}
