+++ integration/vm_test.go
@@ -5,72 +5,21 @@
 package integration
 
 import (
-	"encoding/json"
-	"fmt"
+	"testing"
 	"time"
 
-	"github.com/digitalocean/go-qemu/qmp"
-	"github.com/u-root/u-root/pkg/qemu"
+	"github.com/anatol/vmtest"
 )
 
-type FlashDevice struct {
-	Image string
-}
-
-func (d FlashDevice) Cmdline() []string {
-	return []string{"-drive", "file=" + d.Image + ",format=raw,if=mtd"}
-}
-
-func (d FlashDevice) KArgs() []string {
-	return nil
-}
-
-type MemoryDevice struct {
-	MB int
-}
-
-func (d MemoryDevice) Cmdline() []string {
-	return []string{"-m", fmt.Sprintf("%d", d.MB)}
-}
-
-func (d MemoryDevice) KArgs() []string {
-	return nil
-}
-
-type MachineDevice struct {
-	Board string
-}
-
-func (d MachineDevice) Cmdline() []string {
-	return []string{"-M", d.Board}
-}
-
-func (d MachineDevice) KArgs() []string {
-	return nil
-}
-
-type QemuMonitorDevice struct {
-	Socket string
-}
-
-func (d QemuMonitorDevice) Cmdline() []string {
-	return []string{"-qmp", "unix:" + d.Socket + ",server,nowait"}
-}
-
-func (d QemuMonitorDevice) KArgs() []string {
-	return nil
-}
-
-type VirtioRngDevice struct {
-}
-
-func (d VirtioRngDevice) Cmdline() []string {
-	return []string{"-device", "virtio-rng-pci"}
-}
-
-func (d VirtioRngDevice) KArgs() []string {
-	return nil
-}
+var (
+	machine = []string{"-machine", "virt"}
+	cpu32   = []string{"-cpu", "cortex-a15"}
+	cpu64   = []string{"-cpu", "cortex-a72"}
+	mem     = []string{"-m", "512"}
+	blkdev  = []string{"-drive", "file=../build/img/rootfs.img,format=raw,if=virtio"}
+	nic     = []string{"-nic", "user,hostfwd=udp::6053-:53,hostfwd=tcp::6443-:443,hostfwd=tcp::9370-:9370,model=virtio"}
+	rng     = []string{"-device", "virtio-rng"}
+)
 
 type QOMInteger struct {
 	Path     string `json:"path"`
@@ -80,41 +29,53 @@ type QOMInteger struct {
 
 type TestVM struct {
 	cleanup func()
-	m       qmp.Monitor
-	qemu.VM
+	vmtest.Qemu
 }
 
-func NewTestVM(vm *qemu.VM, monsock string, cleanup func()) (*TestVM, error) {
-	mon, err := qmp.NewSocketMonitor("unix", monsock, 2*time.Second)
-	if err != nil {
-		return nil, fmt.Errorf("Failed to attach to QMP monitor: %v", err)
+func cmdline(args ...[]string) (cmdline []string) {
+	for _, arg := range args {
+		cmdline = append(cmdline, arg...)
 	}
+	return
+}
 
-	if err := mon.Connect(); err != nil {
+func NewTestVM(t *testing.T, bit int, timeout time.Duration, cleanup func()) (*TestVM, error) {
+	var cpu []string
+	var arch string
+	switch bit {
+	case 32:
+		cpu = cpu32
+		arch = "arm"
+	case 64:
+		cpu = cpu64
+		arch = "aarch64"
+	default:
+		t.Fatalf("invalid arch: %dbit", bit)
+	}
+	opts := vmtest.QemuOptions{
+		OperatingSystem: vmtest.OS_OTHER,
+		Architecture:    vmtest.QemuArchitecture(arch),
+		Kernel:          "../build/linux/zImage.boot",
+		Params: cmdline(
+			machine,
+			cpu,
+			mem,
+			blkdev,
+			nic,
+			rng,
+		),
+		Verbose: true,
+		Timeout: timeout,
+	}
+	vm, err := vmtest.NewQemu(&opts)
+	if err != nil {
 		return nil, err
 	}
 
-	return &TestVM{cleanup, mon, *vm}, nil
+	return &TestVM{cleanup, *vm}, nil
 }
 
 func (v TestVM) Close() {
-	v.m.Disconnect()
+	v.Qemu.Kill()
 	v.cleanup()
-}
-
-func (v TestVM) SetQOMInteger(path, property string, value int) error {
-	cmd := qmp.Command{
-		Execute: "qom-set",
-		Args: QOMInteger{
-			Path:     path,
-			Property: property,
-			Value:    value,
-		},
-	}
-	d, err := json.Marshal(cmd)
-	if err != nil {
-		return err
-	}
-	_, err = v.m.Run(d)
-	return err
 }
